% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_col.R
\name{bivar_col}
\alias{bivar_col}
\title{Preview mixed color palette}
\usage{
bivar_col(xcolor1, xcolor2, ycolor1, ycolor2, xnum, ynum, sym = T)
}
\arguments{
\item{xcolor1}{First color for the x variable}

\item{xcolor2}{Last color for the x variable}

\item{ycolor1}{First color for the y variable}

\item{ycolor2}{Last color for the y variable}

\item{xnum}{Number of x categories}

\item{ynum}{Number of y categories}

\item{sym}{Symmetric categories; if true, number of x categories is equal to y categories}
}
\value{
Plot object with the mixed color palettes
}
\description{
This creates a plot that previews your color palette of choice. This also creates part of the legend.
}
