% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_map.R
\name{bivar_map}
\alias{bivar_map}
\title{Create a bivariate choropleth map}
\usage{
bivar_map(
  data,
  xcol,
  ycol,
  xnum,
  ynum,
  title,
  xtitle,
  ytitle,
  xcolor1,
  xcolor2,
  ycolor1,
  ycolor2,
  xstyle = "quantile",
  ystyle = "quantile",
  xbreaks,
  ybreaks,
  legend_pos = c(0.2, 0.2),
  legend_size = c(0.3, 0.3),
  map_style = "choropleth",
  sym = T
)
}
\arguments{
\item{data}{\code{sf} or \code{sp} object}

\item{xcol}{x column}

\item{ycol}{y column}

\item{xnum}{number of x categories}

\item{ynum}{number of y categories}

\item{title}{plot title}

\item{xtitle}{x axis title}

\item{ytitle}{y axis title}

\item{xcolor1}{First color for the x variable}

\item{xcolor2}{Last color for the x variable}

\item{ycolor1}{First color for the y variable}

\item{ycolor2}{Last color for the y variable}

\item{xstyle}{style of splitting x into categories; see \code{classIntervals()}}

\item{ystyle}{style of splitting y into categories; see \code{classIntervals()}}

\item{xbreaks}{optional breaks for x; only needed if xstyle is fixed}

\item{ybreaks}{optional breaks for y; only needed if ystyle is fixed}

\item{legend_pos}{vector of legend position: \code{c(x, y)}}

\item{legend_size}{vector of legend size: \code{c(width, height)}}

\item{map_style}{style of map: choropleth; line, point, etc. to come.}

\item{sym}{Symmetric categories; if true, number of x categories is equal to y categories and x style is equal to y style}
}
\value{
sf object with categorical columns
}
\description{
This function creates a bivariate choropleth map.
}
