% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_pal.R
\name{bivar_pal}
\alias{bivar_pal}
\title{Obtain a color palette from the mixing process}
\usage{
bivar_pal(xcolor1, xcolor2, ycolor1, ycolor2, xnum, ynum, sym = T)
}
\arguments{
\item{xcolor1}{First color for the x variable}

\item{xcolor2}{Last color for the x variable}

\item{ycolor1}{First color for the y variable}

\item{ycolor2}{Last color for the y variable}

\item{xnum}{Number of x categories}

\item{ynum}{Number of y categories}

\item{sym}{Symmetric categories; if true, number of x categories is equal to y categories}
}
\value{
Vector of colors
}
\description{
Creates a color palette from \code{bivar_col()}.
}
