% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val2col.R
\name{val2col}
\alias{val2col}
\title{Value to Color}
\usage{
val2col(z, zlim, col = heat.colors(12), breaks)
}
\arguments{
\item{z}{Vector of numbers}

\item{col}{Vector of colors}
}
\value{
Color levels for each value
}
\description{
This function assigns a color to each value.
}
